requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.4
django-environ==0.4.5
djangorestframework==3.10.2
gunicorn==19.9.0
idna==2.8
mongoengine==0.18.2
pep8==1.7.1
pymongo==3.9.0
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
whitenoise==4.1.3
